헤더 컴포넌트 분리 & 푸터바 틀 완성
