+++ App.js
@@ -5,11 +5,11 @@ import { createStackNavigator } from '@react-navigation/stack';
 import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
 
 // 하단 탭에 들어갈 컴포넌트들
-import FirstStackNavigator from './components/AppTabNavigator/FirstTab.js';
-import SecondStackScreen from "./components/AppTabNavigator/SecondTab.js";
-import ThirdStackScreen from "./components/AppTabNavigator/ThirdTab.js";
-import FourthStackScreen from "./components/AppTabNavigator/fourthTab.js";
-import FifthStackScreen from "./components/AppTabNavigator/FifthTab.js";
+import FirstStackNavigator from './components/tabs/FirstTab.js';
+import SecondStackScreen from "./components/tabs/SecondTab.js";
+import ThirdStackScreen from "./components/tabs/ThirdTab.js";
+import FourthStackScreen from "./components/tabs/fourthTab.js";
+import FifthStackScreen from "./components/tabs/FifthTab.js";
 
 
 // App()
@@ -25,45 +25,69 @@ export default function App() {
 
 
 
-// 해더
-function HeaderBar() {
-	return (
-
-		<View style={{
-			flex: 1,
-			justifyContent: 'space-between',
-			alignContent: "center",
-			flexDirection: 'row',
-		}}>
-			<Image
-				style={{ width: 32, height: 32 }}
-				source={require('./images/image6.png')}
-			/>
-			<Text>asdfasdfasdf</Text>
-			<TouchableOpacity onPress={() => alert('This is a button!')}>
-				<Image
-					style={{ width: 32, height: 32 }}
-					source={require('./images/image6.png')}
-				/>
-			</TouchableOpacity>
-		</View>
-	);
-}
-
-
-
-
 // 하단 네비게이션
 const Tab = createBottomTabNavigator();
 
 function MyTabs() {
 	return (
 		<Tab.Navigator>
-			<Tab.Screen name="First" component={FirstStackNavigator} />
-			<Tab.Screen name="Second" component={SecondStackScreen} />
-			<Tab.Screen name="Third" component={ThirdStackScreen} />
-			<Tab.Screen name="Fourth" component={FourthStackScreen} />
-			<Tab.Screen name="Fifth" component={FifthStackScreen} />
+			<Tab.Screen
+				name="First"
+				component={FirstStackNavigator}
+				options={{
+					tabBarLabel: 'First',
+					tabBarIcon: ({ color, size }) => (
+						<Image
+							style={{ width: 24, height: 24 }}
+							source={require('./images/image6.png')}
+						/>
+					),
+				}}
+			/>
+			<Tab.Screen name="Second" component={SecondStackScreen}
+				options={{
+					tabBarLabel: 'Second',
+					tabBarIcon: ({ color, size }) => (
+						<Image
+							style={{ width: 24, height: 24 }}
+							source={require('./images/image6.png')}
+						/>
+					),
+				}}
+			/>
+			<Tab.Screen name="Third" component={ThirdStackScreen}
+				options={{
+					tabBarLabel: 'Third',
+					tabBarIcon: ({ color, size }) => (
+						<Image
+							style={{ width: 24, height: 24 }}
+							source={require('./images/image6.png')}
+						/>
+					),
+				}}
+			/>
+			<Tab.Screen name="Fourth" component={FourthStackScreen}
+				options={{
+					tabBarLabel: 'Fourth',
+					tabBarIcon: ({ color, size }) => (
+						<Image
+							style={{ width: 24, height: 24 }}
+							source={require('./images/image6.png')}
+						/>
+					),
+				}}
+			/>
+			<Tab.Screen name="Fifth" component={FifthStackScreen}
+				options={{
+					tabBarLabel: 'Fifth',
+					tabBarIcon: ({ color, size }) => (
+						<Image
+							style={{ width: 24, height: 24 }}
+							source={require('./images/image6.png')}
+						/>
+					),
+				}}
+			/>
 		</Tab.Navigator>
 	);
 }
