+++ components/tabs/FirstTab.js
@@ -1,7 +1,7 @@
 import React from 'react';
-import { Text, StyleSheet, View, Button } from 'react-native';
-// import { NavigationContainer } from '@react-navigation/native';
+import { Text, StyleSheet, View, Button, Image, TouchableOpacity } from 'react-native';
 import { createStackNavigator } from '@react-navigation/stack';
+import HeaderBar from "../Header.js";
 // import { SearchBar } from 'react-native-elements';   // 검색바 관련 모듈
 
 
@@ -9,7 +9,13 @@ const FirstStack = createStackNavigator();
 function FirstStackNavigator() {
     return (
         <FirstStack.Navigator>
-            <FirstStack.Screen name="First" component={FirstScreen} />
+            <FirstStack.Screen
+                name="First"
+                component={FirstScreen}
+                options={{
+                    headerTitle: props => <HeaderBar {...props} />
+                }}
+            />
             <FirstStack.Screen name="Details" component={DetailsScreen} />
         </FirstStack.Navigator>
     );
