+++ components/tabs/ThirdTab.js
@@ -1,26 +1,32 @@
 import React from 'react';
 import { Text, StyleSheet, View, Button } from 'react-native';
-// import { NavigationContainer } from '@react-navigation/native';
 import { createStackNavigator } from '@react-navigation/stack';
+import HeaderBar from "../Header.js";
 // import { SearchBar } from 'react-native-elements';   // 검색바 관련 모듈
 
 
-const FifthStack = createStackNavigator();
-function FifthStackScreen() {
+const ThirdStack = createStackNavigator();
+function ThirdStackScreen() {
 	return (
-		<FifthStack.Navigator>
-			<FifthStack.Screen name="Fifth" component={FifthScreen} />
-			<FifthStack.Screen name="Details" component={DetailsScreen} />
-		</FifthStack.Navigator>
+		<ThirdStack.Navigator>
+			<ThirdStack.Screen
+				name="Third"
+				component={ThirdScreen}
+				options={{
+					headerTitle: props => <HeaderBar {...props} />
+				}}
+			/>
+			<ThirdStack.Screen name="Details" component={DetailsScreen} />
+		</ThirdStack.Navigator>
 	);
 }
 
 
 
-function FifthScreen({ navigation }) {
+function ThirdScreen({ navigation }) {
 	return (
 		<View style={{ flex: 1, justifyContent: 'center', alignItems: 'center' }}>
-			<Text>Fifth screen</Text>
+			<Text>Third screen</Text>
 			<Button
 				title="Go to Details"
 				onPress={() => navigation.navigate('Details')}
@@ -42,4 +48,4 @@ const styles = StyleSheet.create({
 
 });
 
-export default FifthStackScreen;
+export default ThirdStackScreen;
