+++ components/tabs/fourthTab.js
@@ -1,26 +1,32 @@
 import React from 'react';
 import { Text, StyleSheet, View, Button } from 'react-native';
-// import { NavigationContainer } from '@react-navigation/native';
 import { createStackNavigator } from '@react-navigation/stack';
+import HeaderBar from "../Header.js";
 // import { SearchBar } from 'react-native-elements';   // 검색바 관련 모듈
 
 
-const ThirdStack = createStackNavigator();
-function ThirdStackScreen() {
+const FourthStack = createStackNavigator();
+function FourthStackScreen() {
 	return (
-		<ThirdStack.Navigator>
-			<ThirdStack.Screen name="Third" component={ThirdScreen} />
-			<ThirdStack.Screen name="Details" component={DetailsScreen} />
-		</ThirdStack.Navigator>
+		<FourthStack.Navigator>
+			<FourthStack.Screen
+				name="Fourth"
+				component={FourthScreen}
+				options={{
+					headerTitle: props => <HeaderBar {...props} />
+				}}
+			/>
+			<FourthStack.Screen name="Details" component={DetailsScreen} />
+		</FourthStack.Navigator>
 	);
 }
 
 
 
-function ThirdScreen({ navigation }) {
+function FourthScreen({ navigation }) {
 	return (
 		<View style={{ flex: 1, justifyContent: 'center', alignItems: 'center' }}>
-			<Text>Third screen</Text>
+			<Text>Fourth screen</Text>
 			<Button
 				title="Go to Details"
 				onPress={() => navigation.navigate('Details')}
@@ -42,4 +48,4 @@ const styles = StyleSheet.create({
 
 });
 
-export default ThirdStackScreen;
+export default FourthStackScreen;
